Reviewed the code

+++ year-2/ray-casting/sphere.hpp
@@ -8,7 +8,7 @@ public:
     Vec3d center;
     double radius;
     int mirroring;
-    Vec3f color = {1, 1, 1};
+    Vec3d color = {1, 1, 1};
 
     Sphere(const Vec3d& center, double radius): center(center), radius(radius), mirroring(5) {}
 };
